+++ appd-pipeline/src/rt/processors/relativeTo.ts
@@ -9,8 +9,8 @@ export class RelativeToProcessor extends BaseProcessor {
         // TODO we could combine all the paths and do a single search, but for now just
         // fire off a search per series
         const args = (node as CommandExpressionNode).args;
-        const path = args.path as string;
-        return Promise.all(
+        const paths = (args.path as string)?.split(';');
+        return Promise.all(paths.map(path => Promise.all(
             flattenDeep(ctx.groups.map(g => 
                 g.map(ts => 
                     ts.sources.map(s => {
@@ -22,8 +22,8 @@ export class RelativeToProcessor extends BaseProcessor {
                         return x;
                     })
                 )
-            ))
-        )
+            )))
+        ))
         .then(flattenDeep)
         .then(results => {
             const mode = args.mode as string;
@@ -43,6 +43,8 @@ function resolveRelative(base:string[], path:string) {
     paths.forEach(p => {
         if (p == '..') {
             r.pop();
+        } else if (p == '.') {
+            // ignore
         } else {
             r.push(p);
         }